General - Добавлен BottomSheet настройки папки - Добавлена возможность выбора цвета

+++ app/src/main/java/com/example/appcash/data/vo/FinanceCategoryVO.kt
@@ -1,8 +1,6 @@
 package com.example.appcash.data.vo
 
-import androidx.compose.ui.graphics.Color
-
 data class FinanceCategoryVO(
     val name: String,
-    val color: Int
+    val color: Long
 )
